Make Field an internal class of Legate store
